Map structure built
Map data stored in JavaScript 2dim array. Each col index in a particular
row currently holds an object. This object has basic information
regarding what the player will see when facing any direction in the
square. This is being created in JavaScript, however the data is stored
in a table and a quesry should be written to utilize the data and
include it in the returned json from the initial ajax call.