DB Class Added
DB class added for use. Initial testing after merging test branch in to
this one resulted in a fail. Issue was any non alpha/numeric characters
used in the ini file must be surrounded by “’s. Some values feel in to
this category. After making the suggested changes, DB class worked as
expected. The solution was fun in the notes section of this page:
http://php.net/manual/en/function.parse-ini-file.php